// types/note.ts
export interface Note {
  id: string;
  title: string;
  content: string;
  createdAt: string;
  updatedAt: string;
  tag: NoteTag;
}

export type NoteTag = 'Todo' | 'Work' | 'Personal' | 'Meeting' | 'Shopping';

export const tagsList: string[] = ['All', 'Todo', 'Work', 'Personal', 'Meeting', 'Shopping'];

export interface CreatedNote {
  title: string;
  content: string;
  tag: NoteTag;
}