// components/NoteForm/NoteForm.tsx
'use client';

import css from './NoteForm.module.css';
import type { CreatedNote, NoteTag } from '../../types/note';
import { useMutation, useQueryClient } from '@tanstack/react-query';
import toast from 'react-hot-toast';

import { useRouter } from 'next/navigation';
import { useNoteDraftStore } from '@/lib/store/noteStore';
import { createNote } from '@/lib/api/clientApi';
import { useDebouncedCallback } from 'use-debounce';

interface NoteFormValues {
  title: string;
  content: string;
  tag: NoteTag;
}

function NoteForm() {
  const router = useRouter();
  const { draft, setDraft, clearDraft } = useNoteDraftStore();
  const queryClient = useQueryClient();

  const { mutate } = useMutation({
    mutationFn: (value: NoteFormValues) => createNote(value),
    onError: () => toast.error('Something went wrong'),
    onSuccess: () => (
      clearDraft(),
      queryClient.invalidateQueries({ queryKey: ['notes'] }),
      router.push('/notes/filter/All')
    ),
  });

  const handleCancel = () => router.push('/notes/filter/All');

  const handleSubmit = useDebouncedCallback((formData: FormData) => {
    const values = Object.fromEntries(formData) as unknown as CreatedNote;
    mutate(values);
  }, 300);

  const handleChange = (
    event: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement | HTMLSelectElement>
  ) => {
    setDraft({ ...draft, [event.target.name]: event.target.value });
  };

  return (
    <form className={css.form} action={handleSubmit}>
      <div className={css.formGroup}>
        <label htmlFor="title">Title</label>
        <input
          id="title"
          type="text"
          name="title"
          className={css.input}
          defaultValue={draft.title}
          onChange={handleChange}
        />
      </div>

      <div className={css.formGroup}>
        <label htmlFor="content">Content</label>
        <textarea
          id="content"
          name="content"
          rows={8}
          className={css.textarea}
          defaultValue={draft.content}
          onChange={handleChange}
        />
      </div>

      <div className={css.formGroup}>
        <label htmlFor="tag">Tag</label>
        <select
          id="tag"
          name="tag"
          className={css.select}
          defaultValue={draft.tag}
          onChange={handleChange}
        >
          <option value="Todo">Todo</option>
          <option value="Work">Work</option>
          <option value="Personal">Personal</option>
          <option value="Meeting">Meeting</option>
          <option value="Shopping">Shopping</option>
        </select>
      </div>

      <div className={css.actions}>
        <button type="button" className={css.cancelButton} onClick={handleCancel}>
          Cancel
        </button>
        <button type="submit" className={css.submitButton}>
          Create note
        </button>
      </div>
    </form>
  );
}

export default NoteForm;
